refactored library by splitting src and include
