+++ lib/.gitignore
