+++ src/utils.cpp
@@ -1,4 +1,4 @@
-#include "utils.hpp"
+#include <base64/utils.hpp>
 
 const std::string utils::base64_charset = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
